move DashboardGridElementModel and DashboardModelEntity to learnroom module
