+++ apps/server/src/modules/learnroom/repo/index.ts
@@ -3,6 +3,7 @@ export { ColumnBoardNode, ColumnBoardNodeProps } from './mikro-orm/column-board-
 export { ColumnBoardNodeRepo } from './mikro-orm/column-board-node.repo';
 export { CourseMikroOrmRepo } from './mikro-orm/course.repo';
 export { DashboardElementRepo } from './mikro-orm/dashboard-element.repo';
+export { DashboardGridElementModel, DashboardModelEntity } from './mikro-orm/dashboard.model.entity';
 export { DashboardRepo } from './mikro-orm/dashboard.repo';
 export { LegacyBoardElement, LegacyBoardElementType } from './mikro-orm/legacy-board-element.entity';
 export { BoardProps, LegacyBoard } from './mikro-orm/legacy-board.entity';
