+++ apps/server/src/modules/learnroom/repo/mikro-orm/dashboard.model.entity.ts
@@ -11,8 +11,8 @@ import {
 } from '@mikro-orm/core';
 import { ObjectId } from '@mikro-orm/mongodb';
 import { BaseEntityWithTimestamps } from '@shared/domain/entity/base.entity';
-import { Course } from './course.entity';
-import { User } from './user.entity';
+import { Course } from '@shared/domain/entity/course.entity';
+import { User } from '@shared/domain/entity/user.entity';
 
 export interface DashboardGridElementModelProperties {
 	id?: string;
