+++ apps/server/src/modules/management/management.entity.imports.ts
@@ -15,6 +15,10 @@ import {
 	LessonBoardElement,
 	TaskBoardElement,
 } from '@modules/learnroom/repo';
+import {
+	DashboardGridElementModel,
+	DashboardModelEntity,
+} from '@modules/learnroom/repo/mikro-orm/dashboard.model.entity';
 import { SchoolSystemOptionsEntity } from '@modules/legacy-school/entity';
 import { MediaSourceEntity } from '@modules/media-source/entity';
 import { OauthSessionTokenEntity } from '@modules/oauth/entity';
@@ -33,7 +37,6 @@ import { ImportUser } from '@modules/user-import/entity';
 import { MediaUserLicenseEntity, UserLicenseEntity } from '@modules/user-license/entity';
 import { Course } from '@shared/domain/entity/course.entity';
 import { CourseGroup } from '@shared/domain/entity/coursegroup.entity';
-import { DashboardGridElementModel, DashboardModelEntity } from '@shared/domain/entity/dashboard.model.entity';
 import { CountyEmbeddable, FederalStateEntity } from '@shared/domain/entity/federal-state.entity';
 import { LessonEntity } from '@shared/domain/entity/lesson.entity';
 import { Material } from '@shared/domain/entity/materials.entity';
